+++ src/commands/login.js
@@ -2,6 +2,8 @@ const crypto = require('crypto');
 const hash = crypto.createHash('md5');
 const inquirer = require('inquirer');
 const request = require('request-promise');
+const emojic = require('emojic');
+const chalk = require('chalk');
 
 import exitWithError from '../common/exit-with-error';
 import { updateJson } from '../common/json';
@@ -16,7 +18,7 @@ export default async userEmail => {
 
   const validateEmail = value => /.+@.+\..+/.test(value);
 
-  console.log('Log in with your ViewAR account.');
+  console.log(`${emojic.bustInSilhouette} Log in with your ViewAR account.`);
   console.log(
     `If you don't have an account, create one at developer.viewar.com.`
   );
@@ -55,8 +57,12 @@ export default async userEmail => {
 
       return config;
     });
-    console.log(`User ${userName} <${email}> logged in.`);
+    console.log(
+      chalk`${
+        emojic.whiteCheckMark
+      } {green User {bold ${userName} <${email}>} logged in.}`
+    );
   } else {
-    await exitWithError(`Authentication failed! (${response.error})`);
+    await exitWithError(`Authentication failed! (${response.error})`, false);
   }
 };