+++ src/common/check-version.js
@@ -1,6 +1,7 @@
 const chalk = require('chalk');
 const latestVersion = require('latest-version');
 const semver = require('semver');
+const emojic = require('emojic');
 import packageJson from '../../package.json';
 
 export default (timeout = 2000) => {
@@ -11,8 +12,10 @@ export default (timeout = 2000) => {
     latestVersion('viewar-cli').then(version => {
       if (!semver.satisfies(currentVersion, `>=${version}`)) {
         console.log(
-          chalk.red.bold(
-            `\Your installed viewar-cli version (${currentVersion}) is not up to date.\nPlease install the latest version (${version}) from npm.\n\thttps://www.npmjs.com/package/viewar-cli.`
+          chalk.red(
+            `${
+              emojic.exclamation
+            } Your installed viewar-cli version (${currentVersion}) is not up to date.\nPlease install the latest version (${version}) from npm.\n\thttps://www.npmjs.com/package/viewar-cli.`
           )
         );
       }