+++ src/common/exit-with-error.js
@@ -1,12 +1,10 @@
 import logger from '../logger/logger';
 
-const chalk = require('chalk');
-
 /**
  * Has to be async to avoid program continuing while server exception loggin is in progress.
  */
-export default async (message, errorCode = 1) => {
-  console.error(chalk`{red ERROR: ${message}}`);
+export default async (message, showAdvancedHint = true, errorCode = 1) => {
+  logger.logError(message, showAdvancedHint);
   await logger.logErrorToServer(message);
   process.exit(errorCode);
 };