+++ src/logger/logger.js
@@ -1,11 +1,14 @@
-import { readJsonSafe } from '../common/json';
+import { readJsonSafe } from '../common/read-json-safe';
 import { cliConfigPath } from '../common/constants';
 import { getErrorLogUrl } from '../common/urls';
 const fetch = require('node-fetch');
 const { URLSearchParams } = require('url');
+const emojic = require('emojic');
+const chalk = require('chalk');
 
 const createLogger = () => {
   let advancedLogging = false;
+  let serverLogging = false;
   let command = '';
   let args = {};
 
@@ -14,15 +17,23 @@ const createLogger = () => {
     args = newArgs;
   };
 
-  const logAdvanced = (...args) => {
-    if (!advancedLogging) {
+  const logError = (message, showAdvancedHint = true) => {
+    console.error(chalk`{red ${emojic.x} ${message}}`);
+    if (showAdvancedHint && !advancedLogging) {
+      console.log(
+        chalk`{yellow ${emojic.warning}  Run with -l for advanced log output.}`
+      );
+    }
+  };
+
+  /**
+   * Logs errors to ViewAR server. Only enabled if flag '--server' or '-s' is set.
+   */
+  const logErrorToServer = async message => {
+    if (!serverLogging) {
       return;
     }
 
-    console.log(...args);
-  };
-
-  const logErrorToServer = async message => {
     const viewarConfig = await readJsonSafe(cliConfigPath);
     const body = {
       type: 'Exception',
@@ -55,8 +66,11 @@ const createLogger = () => {
     set advancedLogging(newAdvancedLogging) {
       advancedLogging = newAdvancedLogging;
     },
+    set serverLogging(newServerLogging) {
+      serverLogging = newServerLogging;
+    },
     setInfo,
-    logAdvanced,
+    logError,
     logErrorToServer,
   };
 };